requirements.txt
anyio==4.12.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
diffusers==0.35.2
docstring-parser==0.17.0
einops==0.8.1
filelock==3.20.0
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.1.7
idna==3.11
imageio==2.35.1
imageio-ffmpeg==0.5.1
importlib-metadata==8.7.0
jinja2==3.1.6
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
packaging==25.0
pillow==10.4.0
psutil==7.1.3
pygments==2.19.2
pyyaml==6.0.3
regex==2025.11.3
requests==2.32.5
rich==13.9.2
safetensors==0.7.0
setuptools==80.9.0
shellingham==1.5.4
shtab==1.8.0
sympy==1.14.0
torch==2.3.1+cu121
torchaudio==2.3.1+cu121
torchvision==0.18.1+cu121
tqdm==4.67.1
typer-slim==0.20.0
typing-extensions==4.15.0
tyro==0.8.5
urllib3==2.5.0
zipp==3.23.0
